refactor(is/is.util): extracting image map functions to utils/image-map
