+++ src/is/is.util.js
@@ -3,9 +3,6 @@
 // separate testing of this file's functions is not necessary,
 // as they're implicitly tested by way of the consumers
 
-import cssEscape from 'css.escape';
-import getDocument from '../util/get-document';
-
 export function isUserModifyWritable(style) {
   // http://www.w3.org/TR/1999/WD-css3-userint-19990916#user-modify
   // https://github.com/medialize/ally.js/issues/17
@@ -30,23 +27,4 @@ export function isScrollableContainer(element, nodeName) {
   }
 
   return element.offsetHeight < element.scrollHeight || element.offsetWidth < element.scrollWidth;
-}
-
-export function getImageOfArea(element) {
-  const map = element.parentElement;
-
-  if (!map.name || map.nodeName.toLowerCase() !== 'map') {
-    return null;
-  }
-
-  // NOTE: image maps can also be applied to <object> with image content,
-  // but no browser supports this at the moment
-
-  // HTML5 specifies HTMLMapElement.images to be an HTMLCollection of all
-  // <img> and <object> referencing the <map> element, but no browser implements this
-  //   http://www.w3.org/TR/html5/embedded-content-0.html#the-map-element
-  //   https://developer.mozilla.org/en-US/docs/Web/API/HTMLMapElement
-  // the image must be valid and loaded for the map to take effect
-  const _document = getDocument(element);
-  return _document.querySelector('img[usemap="#' + cssEscape(map.name) + '"]') || null;
 }
